(fix): For FPS calculated in devtools, mean is now trimmedMean
In case of Chrome Android, there are some frame times that lie much beyond the standard deviation.
This leads to heavy inconsistancies. While these do cause jank, the overall frametime may not be affected.

This change ignores the top and bottom 10% of the outliers, and calculates the mean of the rest of the frametimes.
To include all frames, other metrics like frame_discrepency or frames_per_sec calculated using BenchmarkImpl* should be used.