Don't run AndroidTracingProbe for Cordova apps
AndroidTracingProbe looks for adb logs that are triggered only for Chrome-Android.
All other Chrome WebView apps, including Cordova apps cannot use that.

In AndroidTracingProbe, the capabilities were picked up from the browser session rather than than what was passed in.
As a result, the lines (that were deleted in this commit) simply passed, causing this to be called even for Cordova apps.

As a result, the logs never show up, waitForLogCat timeout and browser-perf failed.

The check in this commit basically says that ALL Chrome-Android browser have an activity, and that should be the only time when AndroidTracingProbe should be triggered.

See #61